skills/web_summarizer/__init__.py
"""
Web Summarizer skill implementation
"""
from typing import Type, List
from pathlib import Path
import dspy

from skills.core.base import Skill


class WebSummarizerSignature(dspy.Signature):
    """DSPy signature for web summarization"""
    
    content: str = dspy.InputField(desc="Web page content to summarize")
    max_length: int = dspy.InputField(desc="Maximum length of summary in words")
    summary: str = dspy.OutputField(desc="Concise summary of the web content")
    key_points: str = dspy.OutputField(desc="Key points as comma-separated list")


class WebSummarizerModule(dspy.Module):
    """DSPy module for web summarization"""
    
    def __init__(self):
        super().__init__()
        self.prog = dspy.ChainOfThought(WebSummarizerSignature)
    
    def forward(self, url: str, max_length: int = 200) -> dspy.Prediction:
        """
        Summarize web content
        
        Args:
            url: URL of the web page (in real impl, would fetch content)
            max_length: Maximum length of summary in words
            
        Returns:
            DSPy prediction with summary and key_points
        """
        # In a real implementation, this would fetch the URL content
        # For now, we'll simulate with a placeholder
        content = f"Content from {url}"
        
        result = self.prog(content=content, max_length=max_length)
        
        return result


class WebSummarizer(Skill):
    """
    Summarizes web content into concise, actionable insights
    """
    
    def __init__(self, metadata_path: Path = None):
        if metadata_path is None:
            metadata_path = Path(__file__).parent / "skill.yaml"
        super().__init__(metadata_path)
        self._module = WebSummarizerModule()
        self._signature = WebSummarizerSignature
    
    def get_module(self) -> dspy.Module:
        """Return the DSPy Module"""
        return self._module
    
    def get_signature(self) -> Type[dspy.Signature]:
        """Return the DSPy Signature"""
        return self._signature
